# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/photo": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change user photo for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Change user photo",
                "parameters": [
                    {
                        "description": "Data to change photo",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.ChangePhotoURL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo edited",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during photo change",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/user/tournaments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get tournaments for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get tournaments for user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort page by name",
                        "name": "sort_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort page by size",
                        "name": "sort_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournaments of user",
                        "schema": {
                            "$ref": "#/definitions/dtos.TournamentsResponse"
                        }
                    },
                    "400": {
                        "description": "Couldn't get tournaments for specific user",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.ChangePhotoURL": {
            "type": "object",
            "required": [
                "photoURL"
            ],
            "properties": {
                "photoURL": {
                    "type": "string"
                }
            }
        },
...
        "dtos.MessageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dtos.TournamentsResponse": {
            "type": "object",
            "required": [
                "tournamentCount",
                "tournaments"
            ],
            "properties": {
                "tournamentCount": {
                    "type": "integer"
                },
                "tournaments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tournament"
                    }
                }
            }
        },
...
        "models.Tournament": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "timesPlayed": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                }
            }
        }
    }
}